Generated claim hash should be unpredictable
Currently, when `claim_hash` is generated, target email and current
time (with seconds precision) is used to generate SHA1. Unfortunately,
this is very predictable, and if the gem happen to be used to confirm
email,– an attacker can easily guess the claim hash value.

This change switches from hashing the email to generating a truly random
string of an equal length (40 characters).